// src/game/types.ts
export enum Character {
    Duke = 'Duke',
    Assassin = 'Assassin',
    Captain = 'Captain',
    Ambassador = 'Ambassador',
    Contessa = 'Contessa',
    // Reformation
    Inquisitor = 'Inquisitor',
    // Promo Cards
    Jester = 'Jester',
    Bureaucrat = 'Bureaucrat',
    Speculator = 'Speculator',
    Socialist = 'Socialist',
}
// ...
export enum Faction {
    Loyalist = 'Loyalist',
    Reformist = 'Reformist',
}

// ── Dificuldade dos Bots ──────────────────────────────────
export enum BotDifficulty {
    Easy = 'Easy',
    Medium = 'Medium',
    Hard = 'Hard',
}

// ── Card de Influência ────────────────────────────────────
export interface InfluenceCard {
    character: Character;
    isRevealed: boolean;
}

// ── Jogador ───────────────────────────────────────────────
export interface Player {
    id: string;
    name: string;
    coins: number;
    influenceCards: InfluenceCard[];
    isHuman: boolean;
    botDifficulty?: BotDifficulty;
    faction?: Faction;
    isEliminated: boolean;
}